fix: remove unused files
