+++ catalyst_voices/apps/voices/lib/app/view/app.dart
@@ -46,9 +46,6 @@ class _AppState extends State<App> {
       BlocProvider<ProposalsCubit>(
         create: (_) => Dependencies.instance.get<ProposalsCubit>(),
       ),
-      BlocProvider<CampaignInfoCubit>(
-        create: (_) => Dependencies.instance.get<CampaignInfoCubit>(),
-      ),
       BlocProvider<CampaignBuilderCubit>(
         create: (_) => Dependencies.instance.get<CampaignBuilderCubit>(),
       ),
