+++ catalyst_voices/apps/voices/lib/pages/proposals/widgets/proposals_fund_info.dart
@@ -1,4 +1,3 @@
-import 'package:catalyst_voices/pages/proposals/widgets/proposals_campaign_details_button.dart';
 import 'package:catalyst_voices_localization/catalyst_voices_localization.dart';
 import 'package:flutter/material.dart';
 
@@ -24,7 +23,6 @@ class ProposalsFundInfo extends StatelessWidget {
             context.l10n.currentCampaignDescription,
             style: Theme.of(context).textTheme.bodyLarge,
           ),
-          const ProposalsCampaignDetailsButton(),
         ],
       ),
     );
