--- catalyst_voices/packages/internal/catalyst_voices_blocs/lib/src/campaign/info/campaign_info.dart
@@ -1,2 +1,0 @@
-export 'campaign_info_cubit.dart';
-export 'campaign_info_state.dart';
