+++ catalyst_voices/packages/internal/catalyst_voices_blocs/lib/src/catalyst_voices_blocs.dart
@@ -5,7 +5,6 @@ export 'brand/brand.dart';
 export 'campaign/campaign_builder/campaign_builder.dart';
 export 'campaign/campaign_stage/campaign_stage.dart';
 export 'campaign/details/campaign_details.dart';
-export 'campaign/info/campaign_info.dart';
 export 'category/category_detail.dart';
 export 'common/bloc_error_emitter_mixin.dart';
 export 'common/bloc_event_transformers.dart';
